pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "django-classy-settings>=3.0.7",
    "django-environ>=0.11.2",
    "django-taggit>=6.1.0",
    "django-tinymce>=4.1.0",
    "django-valkey[brotli,libvalkey]>=0.1.8",
    "django>=5.1.2",
    "djangorestframework-simplejwt[crypto]>=5.3.1",
    "djangorestframework>=3.15.2",
    "drf-redesign>=0.4.0",
    "drf-spectacular[sidecar]>=0.27.2",
    "pillow-avif-plugin>=1.4.6",
    "pillow>=11.0.0",
    "pip>=24.3.1",
    "psycopg[c,pool]>=3.2.3",
    "sorl-thumbnail>=12.11.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "django-coverage-plugin>=3.1.0",
    "django-debug-toolbar>=4.4.6",
    "django-stubs>=5.1.1",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "ruff>=0.7.2",
    "whitenoise>=6.8.2",
    "pytest-asyncio>=0.24.0",
    "pytest-env>=1.1.5",
    "django-browser-reload>=1.17.0",
    "django-extensions>=3.2.3",
    "werkzeug>=3.1.1",
]

[tool.django-stubs]
django_settings_module="django_project.settings"

[tool.pytest_env]
DJANGO_MODE = "test"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_project.settings"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
plugins = ["django_coverage_plugin"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]
"**/{migrations}/*" = ["F401"]

